screen-orientation/active-lock.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<p id="#fragment"></p>
<iframe id="iframe"></iframe>
<script type="module">
  import { makeCleanup } from "./resources/orientation-utils.js";
  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const orientation = screen.orientation.type.startsWith("portrait")
      ? "landscape"
      : "portrait";
    const p = screen.orientation.lock(orientation);
    window.location.href = "#fragment";
    await p;
  }, "When performing a fragment navigation, the orientation change must not abort");

  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());

    // Get the iframe ready...
    const iframe = document.getElementById("iframe");
    await new Promise((resolve) => {
      iframe.onload = resolve;
      iframe.src = "about:blank";
    });

    // Keep references to things we need before navigating.
    const iframeDoc = iframe.contentDocument;
    const iframeWin = iframe.contentWindow;
    const iframeException = iframeWin.DOMException;
    const iframeOrientation = iframeWin.screen.orientation;

    // start changing the orientation
    await test_driver.bless("request full screen", null, iframeWin);
    await iframeDoc.documentElement.requestFullscreen();
    const orientation = iframeOrientation.type.startsWith("portrait")
      ? "landscape"
      : "portrait";
    const p = iframeOrientation.lock(orientation);

    // Navigate the iframe
    iframe.src = "./resources/empty.html";

    await promise_rejects_dom(t, "AbortError", iframeException, p);
  }, "If the document is navigated, the lock request must to abort");
</script>
